skinnymosnter collision is off. made it so it's not hidden.

+++ Source/No_Bark_Vs/Monsters/Base/Monster.h
@@ -103,7 +103,7 @@ public:
 		bool bisMonsterDead;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Status)
-		float MonsterValue;
+		int MonsterValue;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Status)
 		bool bisScoreAdded;
